+++ translations/it.js
@@ -1,0 +1,33 @@
+/**
+ * Italian translation
+ * =====================
+ *
+ * @author:     Patryk Rzucidlo [@ptkdev] <info@ptkdev.it> (https://ptkdev.it)
+ * @license:    This code and contributions have 'GNU General Public License v3'
+ * @version:    0.1
+ * @changelog:  0.1 initial release
+ *
+ */
+module.exports = {
+	"hello": "Ciao! Benvenuto nel bot telegram di ",
+	"hello_next": "Riceverai gli ultimi post tra pochi minuti e appena saranno pubblicati i nuovi.",
+    "command_help": "aiuto",
+    "command_lastpost": "ultimipost",
+    "command_credits": "crediti",
+    "command_email": "email",
+
+    "command_email_notvalid": "Email non valida, esempio",
+    "command_email_entervalidtext": "Inserisci un testo valido, esempio",
+    "command_email_emailfrom": "BOT: Email da ",
+    "command_email_error": "Errore...",
+    "command_email_entervalidformat": "Riprova usando il formato",
+    "command_email_done": "Email inviata :)",
+
+    "command_credits_text": "Bot telegram ope nsource in GPLv3 creato da @ptkdev",
+    "command_credits_url": "https://github.com/social-manager-tools/wordpress-telegram-bot.js",
+
+    "command_help_text": "/ultimipost - mostra gli ultimi post pubblicati"+
+			 			 "/email - contattaci con un'email"+
+						 "/aiuto - mostra la lista di comandi del bot "+
+			 			 "/crediti - informazioni sul bot"
+}