+++ CHANGELOG.md
@@ -1,3 +1,14 @@
-# v0.1.0 (February 5, 2018)
+# v0.1.0 (February 7, 2018)
 
-* First Release in development
+* First Release
+* Easy to use
+* Multi-Session
+* Multi-Language (Italian and English)
+* Multi-Platform (Windows 10, Mac OSX, Linux)
+* Personalizable
+* Custom commands
+* Email contact form
+* Database with user information
+* No wordpress-plugin required
+* No wordpress-mysql user/pass required
+* Wordpress news feed on telegram bot